Organized the code.
Added a couple of useful functions like an overloaded version of "say" that takes a message and also the amount of time it will be displayed before adding another message. Also I added another wait function that takes time as a parameter if a different wait time than generic is needed.